kinetic_tower/requirements.txt
Adafruit-Blinka                      8.5.0
adafruit-circuitpython-busdevice     5.2.3
adafruit-circuitpython-ina219        3.4.20
adafruit-circuitpython-led-animation 2.5.20
adafruit-circuitpython-lis3dh        5.1.18
adafruit-circuitpython-neopixel      6.3.6
adafruit-circuitpython-pixelbuf      1.1.8
adafruit-circuitpython-register      1.9.14
adafruit-circuitpython-requests      1.12.8
adafruit-circuitpython-typing        1.8.1
Adafruit-GPIO                        1.0.3
Adafruit-PlatformDetect              3.30.0
Adafruit-PureIO                      1.1.9
adafruit-python-shell                1.3.3
args                                 0.1.0
black                                22.10.0
click                                8.1.3
clint                                0.5.1
Colr                                 0.9.1
dicttoxml                            1.7.15
machine                              0.0.1
mock                                 4.0.3
mypy-extensions                      0.4.3
packaging                            21.3
pathspec                             0.10.2
pip                                  20.3.4
pkg-resources                        0.0.0
platformdirs                         2.5.4
ply                                  3.11
pycrypto                             2.6.1
pyftdi                               0.54.0
pyparsing                            3.0.9
pyserial                             3.5
pyusb                                1.2.1
rpi-ws281x                           4.3.4
RPi.GPIO                             0.7.1
setuptools                           44.1.1
sip                                  6.7.5
spidev                               3.5
sysv-ipc                             1.1.0
toml                                 0.10.2
tomli                                2.0.1
typing-extensions                    4.3.0
